requirements.txt
python-telegram-bot==20.7
beanie==1.23.5
motor==3.6.0
pydantic==1.10.13
python-dotenv==1.0.1
dnspython==2.6.1
fastapi==0.95.2
uvicorn==0.23.2
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
certifi==2024.2.2
